Add files via upload
Please check this example of the add_a_link.php and let me know if you think this would be a great spot for the comment sections. Also, i chose to use *~*~*~*~*~ instead of ****** I though it stuck out more and brought the attention to the eye better. If you would prefer me to use ********** I will be more than happy to go back and change it. 

Thank you (: